// src/interfaces/OrderTransaction.ts
export interface iQueryOrderTransaction {
  searchStart: string
  searchEnd: string
  typePlatform: string
  invoiceNumber: string
  userId?: number
  username?: string
  purchaseOrder?: string
  name?: string
  name2?: string
  firstname?: string
  lastname?: string
  profitCenter?: string
  materialProductCode?: string
  accountCode?: number
  salesmanCode?: number
  invoiceDate?: string
  address?: string
  address2?: string
  address3?: string
  postCode?: string
  city?: string
  tel?: string
  requireTaxInvoice?: boolean
  taxId?: string
  itemCat?: string
  quantity?: number
  mg4?: string
  UOM?: string
  storageLocation?: string
  SORPrice?: number
  search?: string
}